+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Examples.hs
@@ -28,6 +28,7 @@ module Test.Cardano.Ledger.Shelley.Examples (
   examplePoolDistr,
   exampleStakePoolParams,
   exampleTxIns,
+  exampleVrfVerKeyHash,
   exampleProposedPPUpdates,
   exampleByronAddress,
   exampleShelleyScript,